eventlogging: use /srv/log instead of /var/log as default logging dir
This change is useful to avoid that the root partition (hosting /var/log)
gets the chance to be filled up when the systemd daemons need to spam
the log files for various reasons (in case of eventlogging analytics,
a schema validation error for example). The root partition is usually tiny
(~40G) and not meant to store a bit amount of log files.

The eventbus role will be upgraded in a separate step.

Change-Id: Ief3d57b70fb6612a8cb89690535d7f4cdda5110b