mediawiki::web::vhost: use handler to do proxying
Instead of writing individual ProxyPass conditions for each url, we can
just go through the normal rewrites and then use a SetHandler directive
to proxy all requests for php files to the fcgi server.

Also, convert the Alias directives we had to RewriteRules as that's what
sets SCRIPT_FILENAME and REQUEST_URI correctly when using
SetHandler. Please consider the Aliases had exactly zero effect before
as they were all covered by the ProxyPass directives above.

This basically is a per-vhost version of the catchall we have in the
general configuration, which uses rewrite rules instead (which should be
less efficient). Also set retry=0 globally for our proxy if the feature
is enabled globally.

Change-Id: I52805278d6b2b89de08aa2498dac418ef4f2efdd