icinga: allow configuring max_concurrent_checks via Hiera
On einsteinium the current value is 10000.

The CPU hardware on icinga1001 is actally slightly less powerful
than the one on einsteinium.

"Tuning Nagios For Maximum Performance", point 7, explains:
https://assets.nagios.com/downloads/nagioscore/docs/nagioscore/3/en/tuning.html

Latency on icinga1001 was over 10, so considered High per this doc, which means
"you are probably starving Nagios of the checks it needs".

Which means we are supposed to increase max_concurrent_checks further.

For testing purposes we manually set it to 0 on icinga1001 and restarted icinga.

This this value is one of the main ones to tune performance, move it out of the
template and into Hiera. Allow us to set it differently per host.

Set it to the existing 10000 on einsteinium, so should be noop in prod.

performance data:

https://icinga.wikimedia.org/cgi-bin/icinga/extinfo.cgi?type=4
vs
https://icinga-stretch.wikimedia.org/cgi-bin/icinga/extinfo.cgi?type=4

Bug: T202782
Change-Id: I13c797c79ff86827ae0dfebc83222cdbae719b1a