sonofgridengine: Make an effort at causing the grid master to bootstrap
The grid master as is runs some kind of init when it installs, but
at that point the NFS isn't likely in place.  You end up with something
that doesn't work.  As a last step in the master profile, this runs the
init script that should bootstrap the DB and so forth correctly (having
done it manually at least once).

Bug: T200557
Change-Id: Ic789c1318ea4360f552a9f097ff26440dc9985dd