ATS: check HTTP responses from prometheus exporter
Not only do we want the prometheus-trafficserver-exporter process to be
running, we also want to ensure that it returns HTTP 200. That does not
happen if the software crashes because of unhandled exceptions, in which
case a 5xx error is returned instead.

With the upgrade to ATS 8.0.0, for instance, -trafficserver-exporter
0.0.2 crashes because of the renaming of some metrics:
https://github.com/apache/trafficserver/pull/3567
The bug is fixed in -trafficserver-exporter 0.2.0.

Add a check ensuring that the exporter returns 200.

Bug: T204232
Bug: T204209
Change-Id: I98031cc9f969a3a6ead31409d32e83499a637506