profile::hadoop::common: raise Yarn zookeeper timeout to 20s
The new hadoop masters, an-master100[12], are currently experiencing
sporadic stalls of the JVM (not related to GC) of 8-10s that force
the Yarn Resource Manager to hit its default zookeeper session timeout
(10s) and failover to the standby. While investigating the issue
(probably related to disk io latencies) I am bumping up the timeout
to allow Yarn to be more resilient.

Bug: T206943
Change-Id: I940900e2d8b4a3239e34f8e4764c1a43365a4ee8