hieradata/labs: remove mysql::server::use_apparmor: false
Which images exactly did Yuvipanda refer to when he commented

"# The images we have for debian don't support it" ?

And is that statement still true? Is mysql::server used?
And if it is, do the images still not support apparmor?

I don't see an obvious use on
https://tools.wmflabs.org/openstack-browser/puppetclass/

The goal is to stop using the mysql module and replace it
with mariadb.

Bug: T162070
Change-Id: I0a58dbb25d4efe6a0402656db4805af196dc007a