wdqs: raise alerting threshold on updater lag for public cluster
As disscussed on the linked phab task, updater lag on the public cluster
isn't as much of an issue as on the internal one. Further more, there
isn't much we can do in the short term when this alert is raised,
which means we are starting to ignore it anyway.

Raising the warning threshold to 6 hours would have prevented most of
the alerts in the last month (not all), so that's probably a good start.

Bug: T199228
Change-Id: I89e7ee3a108d590e27ec08f60cecef13edf3fe14