role::memcached: enable interface::rps for mc1035
We are currently trying to identify the cause of recurrent
timeouts returned by mc1035 to various mcrouters (on mw hosts).
Due to the amount of traffic that the shard handles, it seems
wise to try to squeeze as much performance as possible from
the NIC. This is probably not going to solve the timeout issue,
but it should help mc1035 to handle bursts of traffic.

Bug: T203786
Change-Id: I710dfaaaf5d6f45b06ceddcbff61cfdccf635133