toolforge: factorize some common configs not related to the grid itself
There are some common configs which are not related to the grid itself,
config which may be interesting for servers which are not submit hosts.
Put them in a separate profile for later usage by others.

Note the difference now between the two base profiles:

* profile::toolforge::grid::base
* profile::toolforge::base

The first is the base profile for toolforge grid nodes.
The second (the new one) is the base profile for every instance in the
toolforge cloudvps project, even if they aren't grid nodes.

Change-Id: I711571082e09ee1ef84725d3f4069c5aa5731073
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>