profile::cache::kafka::webrequest: filter out Timestamp:Restart logs
After T27611 Varnish has the capability of trying to fetch .webp
variants of the hottest thumbnails, but this adds extra log entries
to the shared memory log that pollutes webrequests. Every log containing
Timestamp:Restart (and hence not Timestamp:Resp) need to be filtered out,
since it represents a failed fetch for the webp variant (not yet generated
for example) that is followed by a 'regular' fetch for the original image.

Bug: T208752
Change-Id: Ied38e6e2619d698dde1545e4b275f8e79360d3d4