profile::base: allow nodes to page when down
While in most cases our paging alerts are associated with cluster
services, sometimes (esp. for virtualization hosts) the loss of a server
can be critical as it can be the only one hosting paging services, or it
can in itself be disruptive of service. More importantly, the host might
have one service running that would page if not working, but that would
not page if the whole host goes down. So we need a way to make a node
page when it dies, if we need that.
This means that one host's notification can be in three states:
* `enabled` - the normal state, where notifications from this host are
  enabled and the individual host does not page.
* `disabled` - notifications for this host are disabled.
* `critical` - notifications on the host are enabled, and if the host
  goes down, that event will page

A parameter is added to profile::base in order to represent these three
states, in place of $notifications_enabled we had before which passed a
value directly to the monitoring defines.

Change-Id: If4e8bb4f24e681af39fbec15b455f1cd028af096