network::constants labs: move deployment-prep instances out into hiera data
This is so deployment-prep instances don't get trusted all over labs, just
within the project.

There appears to be no way to add new (there are existing ones!) hiera calls to
this module properly (as it can't come from a profile in this case), without
triggering the style lint and can't seem to come via realm.pp.

Change-Id: I625f06d85fb42e6a511c83bf010b0e671459b4c6