cloudvps: neutron: don't restart l3-agent on config changes
Having puppet agent restarting l3-agent can lead to some split-brain situations
we would really like to avoid.
This means that in case of configuration changes to some neutron config files,
manual restart of daemons in cloudnet nodes will be required.
In fact, this manual restart and supervision is not a bad idea.

Bug: T208916
Change-Id: I21fa65bbb03e01d05b1a6891e7ebb7e5793a398b
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>