Make labnodepool1001.eqiad.wmnet a spare system
Nodepool is no more used, the machine can be returned to SRE.
Turn the system into a spare::system.
Drop the old role and associated profiles.
Remove the nodepool module entirely.

Icinga monitoring should be removed as well as part of this patch.

More patches will be done to cleanup the puppet repository, but that
requires actions by several teams.

Bug: T209642
Change-Id: Id33cc54d5acaa402bf8a483bddd8f31e785c8fb4