kubernetes: Move runtime-config to hiera
Hardcoding --runtime-config is not nice and not version indedepent,
which in the kubernetes environment is quite important as the kubernetes
project moves forward fast. Move the configuration in hiera, allowing it
to be supplied by the environment. Do not supply a value for production
as it does not really care for enabling the old hardcoded batch/v2alpha1
API, but do override for tools where it is actually used

Change-Id: If6fd1282e4aad696ce67acd65ac0d3b4a7754d14