ores::redis: Set maxmemory-policy: volatile-lur
The default maxmemory policy in Redis 3.0 is no longer "volatile-lru" as
it used to be in 2.8, but "noeviction". The policies behavior is the
same (but LRU eviction is much more precise in 3.0), so only the default
value changed. Make sure we default to it for the redis 3.0 upgrade in
ORES

Bug: T209628
Change-Id: I8ac53d58737d921539ed62b53fc32044068ec091