icinga: Re-add reload functionality
icinga takes a pretty long time to restart causing issues in the web
interface. On top of that puppet propagated changes are pretty often
exacerbating the issue. However it is possible to reload the service
instead of restarting it, almost fully alieviating this.
We used to avoid this by setting the hasstatus and restart parameters of
the icinga service in puppet, but when doing the stretch migration,
which included a systemd unit file, those service parameters were left
behind. Add them back in

Change-Id: I63c6a7c69b84e64f33841bb329b49e380260b83b