sonofgridengine: Read in horrible configuration output
The SGE output from some commands is in two columns, as one would
expect, but it is separated on a purely visual basis with line continuations
and weird columns that are likely tab-separated, but do we really want to bet on it?
Adding some logic to deal with the way this spits out exec nodes.

Bug: T200557
Change-Id: I43869ea598082b2c79361a13c4b2aa2b64153f46