mariadb: remove section s2 from dbstore2002
Section s2 is on dbstore2002 and dbstore2001, but the backup itself only
used the one from dbstore2001.
On host dbstore2002 we have performance issues (replication of s3
doesn't catches up) because of a failed BBU.
Section s3 has way more schemas than the others, we might have better
performance if we'll have more memory for the instance. This patch will
remove the s2 instance from the host, and gives it's allocated memory
as buffer pool for s3.

Bug: T208320
Change-Id: I8e9b467a0b1999504be58e4ac81d16e017f8d59b