fifo-log-demux: restart upon writer restart, remove cargo cult
Ensure that fifo-log-demux is started/stopped whenever the service that
needs it (trafficserver in our case) is started/stopped.

We already did specify WantedBy, which takes care of the starting part.
Now add Requires for the stopping part, and add After to ensure that
fifo-log-demux is started after the writing service.

Remove SuccessExitStatus=2, cargo-culted from nginx-log-peeker. The
process does not exit with 2 on SIGINT, but rather with 130. There
should be no specific reason for fifo-log-demux to handle SIGINT
gracefully, so get rid of SuccessExitStatus altogether.

Bug: T204225
Change-Id: I30bd486701381c61a358557c3cde5a86b7d6ad4a