mariadb: Modify dump_section to allow different types of dump
This is a first timid modification of dump_section to allow
the usage of mariabackup (mariadb-compiled version of xtrabackup)
when setting type=snapshot instead of the default (type=dump).

This is mostly a prototype so decisions can be made once it works
as indended about separating this on a different utility or
refactoring to allow multiple different types.

Bug: T210292
Change-Id: I11d018294a556d84c9184bf2b53efe4d75cedbdb