tslua: add logic to detect when to run tests
So far, tslua tests have been executed only by accident. Modifications
to tox.ini result in `tox -r` to be executed. By seeing the output in
jenkins I was lead to believe that ts-lua tests were being run when
needed.

However, adding a tox environment to envlist in tox.ini is not enough to
make sure that the relevant tests are executed. Some logic needs to be
added to taskgen too. Do that for ts-lua and mention this new (for me)
discovery in the comment above envlist.

Change-Id: I41632ea50fccd415112df1d95d0b4c2a4ae57626