toolforge: introduce bastion systemd-based resource control
Previously, in bastions, we had resource control based in cgred.
We would like to move forward to a more modern model, using systemd.

Introduce a new approach based on systemd user slices.

This code is only meant to run on Debian Stretch (because dependency on a
concrete systemd feature introduced in v239) and thus only in the Stretch-based
Toolforge version.

Bug: T210098
Change-Id: Id45ee83d57905869f75855b51cc3877996911406
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>