puppet-merge: allow only showing diffs without merging
A lot of our puppet merges need us to coordinate actions across the
cluster, either for canary deployment of puppet changes or whenever we
have to coordinate multiple actions across the cluster (puppet run +
rolling restart of an important service, ensure puppet runs first on one
host and then on the monitoring server, etc).

In order to be able to write spicerack cookbooks to perform these tasks,
we need to be able to launch puppet-merge remotely and non-interactively
as cumin doesn't support remote interactions. We still want to be able
to protect merges of production code from any form of tampering from the
gerrit server, so we allow a two-step non-interactive workflow that
preserves the same anti-tampering guarantees as the old mechanism:

* `puppet-merge -d` will show the diffs and return the SHA1
   corresponding to the commit
* `puppet-merge -q -y <sha1>` This will proceed to non-interactively
  merge the code at the desired SHA1

Change-Id: I48bb405b068416aa9d846367cc24b8af90a7ea8b