icinga: if on passive host, replace icinga-downtime script with warning
As described in the ticket, sometimes users go to an inactive icinga host
to run the icinga-downtime script. It does not display a warning but users
are also not getting a downtime, not realizing they were on the wrong host.

To prevent this, use the existing $is_passive and replace the content of the
script with a warning. Puppet already knows this so this instead of
trying to detect active host within the script itself.

Also move the existing file from icinga module to profile::icinga and
use files/templates in profile module accordingly.

As opposed to adding the is_passive parameter to the icinga class itself.

That would also work but i guess this script belongs more into profile
than the module itself.

Bug: T210380
Change-Id: Ie101e7913d9ac533f02952f7ae17cc838d5b8c91