Gerrit: Fix placement footer in edit view
For whatever reason the header in the edit view is set to 16px. This
causes the body content to be visible, but the actual bounding box is
cut-off. The result is that the footer appears directly under (and very
close to) the close and save buttons in the edit view.

We may need to revisit this css for newer versions.

Change-Id: I726f927c18d8d74859e9668487a03b860ad1d964