peopleweb: allow deployment server to connect to port 80
Currently only caching servers are allowed to talk to port 80
to the backend(s) of peopleweb.

But that means we can't test for changes with
apache-fast-test from where the test script is installed.
(deployment_servers, cluster::management, not cache servers).

Allowing deployment server (IPv4 and IPv6) to connect to 80 for
testing.

Abandoned another approach in the past to solve this by adding
the test script on cache servers (I8b5f56381a456366), so now this.

Deliberately using a second ferm::service stanza instead of combining
it into a single srange. Seemed the cleanest way to me to have a separate
snippet in ferm/conf.d and also most readable here in puppet. No escaping
of $CACHES to get around puppet-lint etc.

Bug: T210036
Change-Id: If175edf56d1ad451529a9cdbb1849a8401bdf493