sonofgridengine: correct shadow and master init issues
Standing up a new grid master did strange things because the service
statement predated systemd.  While systemd is using a sysv script, it
still has a pid file and maintains service info.  Puppet does bad things
without that like starting a root grid that has no access to anything.

The shadow needs an init, so this systemd might actually work with the
wrapper script it uses.   Also, moving the software installs below the
directory setup in puppet so that things install in the correct fashion.

Bug: T200557
Change-Id: I87a4d936aee8ce293b3925b97bc75588190c4c83