EventLogging Logstash filter: move useful fields out of event
ElasticSearch does not handle mixed-type fields (fields which are
objects in some documents and not objects in others within the same
index) well, and event is a super common name that's bound to be
used as a string somewhere, so something like event.schema cannot
be used in saved searches and visualizations. Move them one level up
(and override the fairly useless EventError schema properties there).

Bug: T205437
Change-Id: Ieda875ec48213fdeffb379514e63e131704363cf