service::node: add the 'use_nodejs10' parameter
The idea is to centralize the declaration of the nodejs10
component and allow all the users of service::node to simply
enable/disable it via a parameter (set in various profiles
for example).

Removing also the require_package function to avoid a
dependency cycle, as far as I can see we don't really need
it but I might be wrong.

Bug: T210704
Change-Id: I8a8f0d5177100183fd925fa656887574261b1140