toolforge: homogenize system::role calls
Different roles call system::role in differents way.
Also, some roles use a shortened version of the role name (without the 'role::
string).

Address all this by switching to use the $name variable and introducing a
common pattern in all the calls.

This is just a cosmetic change for preserving a common format in our Toolforge
roles.

Change-Id: If05ac1b598ea3149f792fbaca23836a4e14e9292
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>