switch graphite host for prod cassandra from graphite1003 to 'none'
site.pp says graphite1004 is now the primary graphite host
and graphite1003 is a spare system.

But here graphite1003 is still set as the host for the prod
cassandra cluster.

Per previous comments set to a bogus value instead of the new
server.  Adding FIXME comment to make clear no server is used.

Bug: T209357
Change-Id: I03c8bb747b80faacb9ffe70fb1c7a56dbd26f839