contint: rephrase comment for castor rsync server
We disallow DNS some DNS lookup on the Castor rsync server. We do not
need the feature and previously it caused bunch of log spam due to DNS
issues on WMCS.
Just rephrase the comment to drop the obsolete reference to
contintcloud. We still do not need the DNS lookups.

Change-Id: I3de33225c9d419296db43873c9d6629587d56973