gridengine: simplifying the config and making it more "normal" for grid
This is somehow failing to correctly update configs for shadow, so
I'm making this more like a typical grid launch rather than debugging mode.
It's a thought.  Also, since grid does a lot of games with Tmp/Home and
friends, and grid masters require login to use, removing the restrictions
on the process.

Bug: T211258
Change-Id: Ibf49fe1c748582ca6f9cac5300cbb4ae8751a252