kernel.mtail needs hostnames
Right now the only place that kernel.mtail runs is on the
syslog::centralserver hosts.  Because we don't parse originating
hostname out of the logs, they don't presently make it to Prometheus,
which means that we wind up with a single timeseries per metric per
centralserver, which is not really useful.  For instance, here is the
complete list of cpu_throttled timeseries we have in eqiad:

cpu_throttled{cluster="misc",instance="centrallog1001:3903",job="mtail",
              prog="kernel.mtail",site="eqiad"}

cpu_throttled{cluster="misc",instance="lithium:3903",job="mtail",
              prog="kernel.mtail",site="eqiad"}

With no way to tie it back to originating machine, these are not all
that useful.

This change adds hostnames from syslog for all kernel.mtail metrics
and adds tests for metrics that lacked them.

Bug: T209863
Change-Id: I4c01342d41e0c6fa9989c5f8503589e213b569ac