Nova: lower cpu_allocation_ratio by a lot
The default value for this is 16.  I can see the rationale for
this but given that we do sometimes see hypervisors maxing
out their CPU usuage I think a more modest number is in order.

There's no solid justification for 4.0 other than that it
might at some point actually affect scheduling, which 16
definitely will not.

Change-Id: I90954e2765ace797ec318be5e0ecdd14243be326