role::beta: introduce docker_services
In this patch we introduce a simple system to allow to run services from
docker images in deployment-prep.

If you want to run a service from a docker image in beta, you just need
to apply this role to the VM you chose, and to configure its hiera
variables.

This is a temporary solution while we evaluate more options.

Change-Id: I884aaf73280eca873131e654cf78d5e0476233e1