authdns config/CI refactor [5/5]
Update authdns-local-update to use the authdns-gen-zones.py that's
shipped by the ops/dns repo, and get rid of the one we were
shipping from puppet to /usr/local/bin.

Replaces authdns-lint with a simple wrapper to execute the ops/dns
repo's check-gdnsd.sh correctly.  We can get rid of authdns-lint
entirely after we've migrated CI to a stretch-based Dockerfile.

Bug: T205439
Change-Id: I9f39c644f66ac780633a7754626b35cd08830df6