mediawiki/scap: do not install sql scripts on canary appservers
A canary appserver does not have a mysql client installed,
a mw-maintenance server does.

We should not install the scap "sql" scripts if there is no
mysql client that they require as the ticket describes.

Restructure the setup so that canary appservers don't get
these 2 scripts when including the scap scripts class.

But avoid having to check for "is fqdn like maintenance server
in Hiera" or similar.

Bug: T211512
Change-Id: Ifa340fdd8e31731e512bff19817504c0ed8d468a