Swap oxygen with weblog1001
The procedure should look like the following:
1) disable puppet on oxygen/weblog1001 and merge this change
2) run puppet on oxygen and mask all the kafkatee instances
3) run puppet on weblog1001 and spin up the new kafkatee instances,
   that should start kafka consumer clients from the latest offset.

This means that for a brief moment we will not have 503 logs flowing
to logstash and on disk, but it should be fine.

Oxygen will be kept around for a bit in case the sampled webrequest log
is needed, then it will be properly decommissioned.

Bug: T211883
Change-Id: I56c0ff391c62351f3c0425b3e0dfdef45c9e5ac9