cloudvps: make spreadcheck monitor region-aware
Give region support to the script and config. The script doesn't need any
special modification since the config parameter 'region_name' can be passed
directly to the novaclient constructor.

Introduce a new hiera key to allow running the check in an arbitrary region.
This is required because right now Toolforge is running in 'eqiad' and will
be in 'eqiad1-r' later. That's why we can't reuse existing eqiad1 hiera keys
for region specification.

Previous to this patch, the script would throw the exception
'novaclient.exceptions.AmbiguousEndpoints' if running against a keystone
serving several regions, which is our case with eqiad/eqiad1-r.

Bug: T211451
Change-Id: I6127996294e09d9ba8b16b41b06c60faad2dcdf5
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>