apt: add component/pyall to stretch-wikimedia
This is to host backported or forward-ported versions of Python that did
not ship with stretch. Right now these would be Python 3.4, 3.6 and 3.7.

Note that libraries shipped with stretch won't necessarily work with
those Python versions, either because the code is incompatible (e.g. use
of the async keyword) or because the package in stretch included a
native extension. As such, these packages will primarily be useful for
creating virtual environments without the system's site-packages.

Change-Id: Ie229316a152e14febc9fe7ed96fd3ceb4e20b17e