Add eevans, mobrovac, ppchelko to proton-admins
The 3 users used to have access to the service via the use of the
sc-admins, which was inherited from the sca/scb roles when instantiating
the cluster and the host. As such, the sc-admins group was temporary and
was meant to be used only during the handoff period. Now that the
sc-admins group has been removed, add the 3 users of the sc-admins group
to the proton-admins group in order to restore rights.

Change-Id: I20ddc06c1855824f92d6077d7642481f2fe6e2d3