sonofgridengine: restrict the cluster from initializing itself
Just in case there is any chance that this cluster might somehow run
the db init on the wrong NFS share, I'm disabling that function until
the rest of the manifest is manually inspected for quirks before running
on the new SGE grid master in actual ToolForge.

Just trying to prevent disasters.

Bug: T212153
Change-Id: Ie60f736f58a7e4f3fd350cb06f3363d53646c320