Add proxy info to toolkit analyzer cron job
The analyzer toolkit queries[0] wikidata.org for some properties and thus
must go through the datacenter's proxy. Based on code someplace else in
this repo, there seems to be a `site` constant that holds the
datacenter's name.
As far as I can tell, there is no user and password to be used for that
proxy.

[0] https://gerrit.wikimedia.org/r/475807

Bug: T209399
Change-Id: I046ae05347e4f93d4eed4a7895297988d287dca9