sonofgridengine: change back to simple foreground service for the master
This seems to be the most reliable systemd setup for the master end after much
testing. This service doesn't really seem to like systemd much, lol.

For some reason, in this particular form, it isn't doing awful spam in the messages
file and is working right.  I'm going to walk away and leave it.  I don't know
what this will buy us in the end, but I'm trying everything.

Bug: T211258
Change-Id: I7662f6d125502870ce1fb4d3c1bc17e3aa498d1e