profile::mediawiki::php::monitoring: fine-grained opcache invalidation
With this change, the /opcache-free endpoint can accept a path (either a
file or a directory, but basically the argument of either scap sync-file
or scap sync-dir) and invalidate the opcache for all php files
included. This will allow scap to selectively invalidate cached files
during sync-file and sync-dir, thus eliminating the need to nuke all of
our opcache when changing just one file.

Also, fix a bug in dump_file.

Bug: T211964
Change-Id: Iaf9effe0c565097e721a1b748afc5927d139dc0e