git: git::clone requires the git package
When testing a manifest that relies on git::clone, it would fail due to
lack of Package['git']. It is defined in base::standard_package but it
seems over kill to have to include base for a single package.

Add a require_package('git') so we do not need the base module.

Change-Id: I136609aaa87c76cd389e6f68b8a64c83c6f250c6