doc: relocate from /srv to /srv/docroot
integration/docroot.git can't be cloned at /srv since that is owned by
root:root. We have the issue on the old doc host which is the subject of
T149924.

The git clone is done under /srv/docroot, this update Apache, Rsync and
Bacula to use that directory instead.

Also made the backup more restrictive, we only care about the published
doc not the whole tree.

Bug: T137890
Bug: T149924
Change-Id: I09c85a47e907d4fd6ffb52ece893161df44e2bd6