Add kafka reporting topic to Puppet config
Since we have sometimes events going to non-primary cluster, it
creates weird reporting, so we want to set proper reporting topic.

Since T207994 does not seem to be fixed anytime soon (and maybe
doesn't need fixing) it's time to have it properly encoded in
puppet.
See https://phabricator.wikimedia.org/T207994 for details

TODO: maybe we could use the name of the actual active cluster,
so it would switch automatically if clusters switch.

Change-Id: I3d9f7a453461826b443ab0cd3601ef8d20a2a06e