Monitoring: add VRRP check
Add a monitoring check for testing the consistency of a VRRP
configuration against a set of two routers. It's currently only tested
with Juniper routers, but not limited to them per se.

This is written in Python, and uses snimpy and the (slightly complicated
and very controversial) Python3 nagiosplugin module.

Added to faidon's initial CR:
- code to install python3-nagiosplugin
- nagios configuration
- puppet stanza to deploy

The check is applied to only one of the pair of VRRP routers, the 2nd
router's FQDN is passed as parameter "vrrp_peer".

Bug: T150264
Change-Id: I3cd62ccfb621108582474a65bde13fbe00230c1b