toolforge: Redirect tools-static to https
Unconditionally redirect from http to https for the
tools-static.wmflabs.org site. This should be safe as tools-static
serves static content and should not be receiving POST/PUT/PATCH verbs
which are not safe to redirect. Also set an HSTS header to pin to https
for the following 24 hours. We will raise this later, but for now this
gives us an escape mechanism if we find that forcing https creates some
unexpected problems.

Bug: T102367
Change-Id: I9f21c978640def802b9b38130c2d3f9346fb83ad