testreduce: if on stretch, use stretch-backports to get npm package
The npm package wasn't available in stretch but it now is via
stretch-backports per Moritz.

This should finally unblock the replacement of ruthenium (jessie)
with scandium (stretch) which has been stalled for a while.

apt-pinning all the node-* packages as listed in the comments to
make it possible without dependency errors.

Bug: T201366
Change-Id: Id7687528d941c44dd9d748cdf5e7b13ba7a806ff