systemd::syslog: allow to add the 'stop' rule when needed
There are use cases in which daemons can log a lot of useful data,
that should not be replicated to syslog/daemon.log, since it may
cause side effects like the growth of root partition's space used.

In order to avoid any change in the actual logging scheme, a new
parameter is introduced that is a noop by default.

Bug: T212915
Change-Id: I93a5998b0f982dc10e69589a5259714d591bbf21