admin: test for absent users
We recently marked 'ladsgroup' as being absent via ensure: absent but
the user had not been added as a member of the 'absent' group.

Add a test that ensures that members of the group and users marked
absent match.
Slightly refactor the test to open the data.yaml file only once for the
test class.

Ran test via: tox -e admin

Change-Id: I8f17b2786ec034456ef8ea898afefc8294079f9a