Introduce role::analytics_test_cluster
These classes are going to be used temporarily to ease the testing
of Kerberos settings in production. The code is basically the one
running for the Analytics cluster, should be the easiest solution
to avoid any overlapping (expecially via Hiera keys).

Proposed mapping:
- masters: analytics102[8,9]
- coord: analytics1030
- workers: analytics10[31-41]

For analytics-tool100X we could ask for 3 new ganeti VMs.

Bug: T212256
Change-Id: I2b2dff8d06725fee2621a56b208a72a03cb522e4