Enable base::service_auto_restart for passive Icinga node
Automatically restart Apache on the passive node if underlying
runtime library dependencies have been upgraded. For the primary
server the restarts still need to be coordinated to prevent that
Apache is being restarted in the middle of an outage or a similar
unsuitable situation.

Tested on icinga2001 after applying the latest OpenSSL
security update.

Bug: T135991
Change-Id: Ic52b9fe114f47d5bdfc5c4fece02d4665ef81984