create static microsite for RT ticket archive, copy static Bugzilla
Copy the setup of https://static-bugzilla.wikimedia.org/ to create
a similar static microsite as a place to put RT ticket content.

The plan is to put static HTML files scraped from RT as an archive
behind an LDAP login and then be able to deinstall the actual RT
Perl application, decom the jessie host ununpentium and only
keep this static site around. So everything just like we did for
Bugzilla back then. Hence copying the classes with minor adjustments.

Bug: T180641
Change-Id: If83b2b062ff0e71cd3c6459a436f2c1044125f1c