toolforge: refactor docker builder profile
Refresh the puppet code for docker-builder servers in toolforge.
While at it, give support for Debian Stretch, which is mostly working except
a missing declaration of the 'docker' group.

This patch means that the original class:
  role::toollabs::docker::builder
was converted into:
 role::wmcs::toolforge::docker::builder
 profile::toolforge::docker::builder

Since the old role is no longer used, delete the code now.

Bug: T213418
Change-Id: I385f95e9f28a48d854cdbe1b377978e7163cb3d7
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>