testreduce: also pin nodejs,nodejs-dev,npm to stretch-backports
npm and nodejs both need to be installed from stretch-backports,
one can't be without the other.

The nodejs-dev package is needed in addition to nodejs.

Priority needs to be higher than 1002 or due to the existing
pins we get nodejs 6 from wikimedia-backports instead of nodejs 8
from stretch-backports and things break.

Bug: T201366
Change-Id: I50ac8a3112eca07c20a4778c380d323397272d6f