service::node: do not install nodejs-legacy if on stretch
As done in I50ac8a3112eca07c we are installing nodejs
from stretch-backports which now gives us nodejs 8,
as opposed to wikimedia-backports which has nodejs 6.

After having that fixed the next issue is that service::node
attempts to also install nodejs-legacy (6) which now conflicts.

So, if on stretch or newer do not install nodesj-legacy anymore.

Bug: T201366
Change-Id: I15ec60055748b4ce5a2c410602bc6fd778634277