services: add missing 'mediawiki/services' prefix to git cloning
The way this is currenly setup puppet will end up with a commandline:

/usr/bin/git  clone --recurse-submodules \
https://gerrit.wikimedia.org/r/parsoid/deploy \
/srv/deployment/parsoid/deploy

and that fails because the "mediawiki/services" prefix is missing
from the repo name.

It is unclear how this worked before, the code all just uses $title
and that prefix does not seem to be defined anywhere.

The relevant defined type and class all just get "parsoid" as title
and $title/deploy is set as the repo.

Setting it in here should be fine if this defined type is only made
for mediawiki/services.

Bug: T201366
Change-Id: Ieeaaa44ed66e4a88c863fbdffc77b3d1c45c913b