role: add prometheus2 backwards-compatibility rules
prometheus2 changed the rules file format to yaml and provided 'promtool update rules <file>'
command to automatically convert from the old format to the new yaml format.

This changeset includes the result of running that command on our existing rules and adds
the prometheus-node-exporter backwards-compatibility rules.  It takes the new names
and writes them also to the old names.

Bug: T213708
Change-Id: Ied4d6d6a6294cc1b82f3c965e391918ae33216cb