esams/eqsin: flip unified to globalsign-2018
The current Digicert cert expires on Jan 24.  It's Friday and we're
closing in on clock-skew-sensitive ranges by Monday, so flipping to
GlobalSign at the Digicert-using sites today to avoid risks.  Even if we
manage to get a certificate issued today, we'd have to make this
failover for a few days to account for acceptable client clock skew
windows.

Bug: T209515
Bug: T210639
Change-Id: I7960c16a05e18e88553a14ee36088e2e8a9a3dcc