toolforge: move SGE diamond collector to cronrunner on Trusty grid
Restore metrics from the Trusty grid engine deploy by moving the
diamond collector that was historically on the services nodes to the
cronrunner nodes. This will all die in the future when we remove all of
the Trusty nodes.

The same diamond collector is absented from the new
profile::toolforge::services::basic manifest. The services nodes using
the new profile are not connected to the grid and thus can't run this
code properly. This will be addressed in T211684.

Bug: T214182
Change-Id: I7458fc3ef1deff6edeb5f8e29262588f5a46ec36