dbstore_multiinstance: Add staging db
Add the possibility of having the staging database on dbstore
multi-instance host. If this is finally what is needed T210478#4894472
This will be just used for dbstore1003-1005
These hosts have the staging database in a different instance, not
like they do know no dbstore1002 which is mixed with all the other
core databases to allow joins against them.

I would give it port 3340 to make sure it doesn't interfere with
the current core sections.

The buffer pool size is just a test, not sure if it should be 10G or more.
That can be adjusted as we go.

This is a proof of concept and it is work in progress I want to
build a whole host as a test, so we can start getting users to
test it and see what we need and what we are missing.
That's the idea with dbstore1003, get the core instances up and
running + the staging database, and ask some users to start
testing it as soon as we have the grants also defined.

Bug: T210478
Change-Id: I9d24ef4cfecdd085b31bdf78faa5892768a3ba0e